// apps/back/src/shared/lib/interfaces/auth.ts
export interface Payload {
    sub: string;
    email: string;
  }
  
  export interface AccessToken {
    accessToken: string;
  }
  
  export interface RefreshToken {
    refreshToken: string;
  }
  
  export interface Tokens {
    accessToken: string;
    refreshToken: string;
  }
  export interface Payload {
    sub: string;
    email: string;
  }
  
  export interface AccessToken {
    accessToken: string;
  }
  
  export interface RefreshToken {
    refreshToken: string;
  }
  
  export interface Tokens {
    accessToken: string;
    refreshToken: string;
  }
    